Refactoring exercise 7 (#20)
* adding names to cities
* solving problems on Position's distance in hexagonal grids

+++ src/main/kotlin/com/github/caay2000/ttk/pathfinding/PathfindingResult.kt
@@ -4,7 +4,7 @@ import com.github.caay2000.ttk.context.world.domain.Cell
 
 data class PathfindingResult(
     val path: List<Cell>,
-    val cost: Float
+    val cost: Int
 ) {
 
     fun removeFirstCell(): PathfindingResult = copy(path = path.drop(1))
